handshake model added; KeyTransfer UI component added; minor code cleanup.
