+++ Limp/Client/Pages/KeyTransfer.razor
@@ -1,0 +1,15 @@
+﻿<div class="encryption-disclaimer">
+    <div class="encryption-message-box">
+        <div class="encryption-message-header">
+            <h3>Please, wait while we working on security and privacy of this chat</h3>
+        </div>
+        <div class="encryption-message-body">
+            <p>We're now creating a new AES message encryption key.</p>
+            <p>We're now sending your conversant an encrypted message containing the encryption key.</p>
+        </div>
+    </div>
+</div>
+
+@code {
+
+}
