+++ Limp/Server/Utilities/Kafka/IMessageBrokerService.cs
@@ -1,4 +1,4 @@
-﻿using ClientServerCommon.Models;
+﻿using ClientServerCommon.Models.Message;
 
 namespace Limp.Server.Utilities.Kafka
 {
