+++ Limp/Shared/Models/Message/Handshake.cs
@@ -1,0 +1,9 @@
+﻿#nullable disable
+using LimpShared.Encryption;
+
+namespace ClientServerCommon.Models.Message;
+
+public class Handshake
+{
+    public Key OfferedAESKey { get; set; }
+}
