+++ Limp/Shared/Models/Message/MessageType.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientServerCommon.Models.Message;
+
+public enum MessageType
+{
+    UserMessage,
+    Handshake
+}
